Prepare to support containerized tests in Makefile
Get the src/ and test/ files recursively (so we can create an arbitrary
structure of subdirectories there).

Rename 'test' target to 'test-exec', and make it clear in the help
message that 'test-exec' and 'test-data' run the tests locally (for
'test-exec', this is only useful when the system is in FIPS mode).